+++ admin/inc/configuration.php
@@ -6,14 +6,17 @@
  * @subpackage Config
  */
 
-$site_full_name     = 'GetSimple Legacy CMS';
-$site_version_no    = '2024.2-alpha';
-$name_url_clean     = lowercase(str_replace(' ', '-', $site_full_name));
-$ver_no_clean       = str_replace('.', '', $site_version_no);
-$site_link_back_url = 'https://github.com/dimayakovlev/getsimple-legacy-cms';
+define('GSNAME', 'GetSimple Legacy CMS');
+define('GSVERSION', '2024.2-alpha');
+define('GSURL', 'https://github.com/dimayakovlev/getsimple-legacy-cms');
+
+// These variables are deprecated. Use constants GSNAME, GSVERSION and GSURL instead.
+$site_full_name     = GSNAME;
+$site_version_no    = GSVERSION;
+$site_link_back_url = GSURL;
 
 // cookie config
-$cookie_name        = lowercase($name_url_clean) . '_cookie_' . $ver_no_clean; // non-hashed name of cookie
+$cookie_name        = lowercase(str_replace(' ', '-', GSNAME)) . '_cookie_' . str_replace(array('.', '-'), '', GSVERSION); // non-hashed name of cookie
 $cookie_login       = 'index.php'; // login redirect
 $cookie_time        = '10800';     // in seconds, 3 hours
 $cookie_path        = '/';         // cookie path
@@ -21,12 +24,4 @@ $cookie_domain      = '';        // cookie domain
 $cookie_secure      = false;        // cookie secure only
 $cookie_httponly    = true;        // cookie http only
 
-$api_url            = 'http://get-simple.info/api/start/v3.php';
-# $api_timeout        = 800; // time in ms defaults to 500
-# $debugApi           = true;
-
 $cookie_redirect = 'pages.php';
-
-if (!defined('GSNAME')) define('GSNAME', $site_full_name);
-if (!defined('GSVERSION')) define('GSVERSION', $site_version_no);
-if (!defined('GSURL')) define('GSURL', $site_link_back_url);